Brief notation of study material for HashMap
The hash map offers functionality that may facilitate finding chunks. The code that handles TileChunks and Map movement are a good candidate for a control class. The WorldManagerAppState is quite bloated compared to other classes and can use some modulaity (currently feels like skimming an article when analyzing the feature list of this class).